update course_design topic2
