+++ course_design/ljh_24_topic2/readme.txt
@@ -1,3 +1,3 @@
 本代码（选题2）在数字图像处理实验环境（ros20.04+opencv）环境下可运行。
-参考之余，可尝试对图像识别速度进行进一步优化。
+代码仅供参考，其中/*以此方式注释*/的部分请自行完成，可尝试对图像识别速度进行进一步优化。
 未参考助教给出的vs版本图像金字塔。
